mvp functionalities and implementations

+++ backend/Cabin-api/src/main/java/com/uef/cabinapi/repository/CabinRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.uef.cabinapi.model.Cabin;
 
-public interface CabinRepository extends JpaRepository<Cabin, Integer> {  
+public interface CabinRepository extends JpaRepository<Cabin, Long> {
 }
